feat(zsh): dump and use completion cache from /tmp
When zsh is started, it needs to make completion cache which will be
used for fast real-time completions. To speedup this process, after
creating it once, it will dump it in the /tmp folder so that next
invocations of zsh can just read it instead of creating it again.
Because the /tmp will be cleaned after every computer restart, it is
safe to also add -C flag which will prevent zsh from comparing the read
completion cache with the available completions on the system and thus
additionally increase its start time, relying on the fact that newest
completions added to the system will be used after computer restart.